refactor: utils 함수별로 파일 분리
