+++ src/shared/_utils/convertHEXToRGB.ts
@@ -1,4 +1,4 @@
-export const convertHEXToRGB = (hex: string) => {
+const convertHEXToRGB = (hex: string) => {
   let hexCode = hex
   if (hexCode.startsWith('#')) {
     hexCode = hexCode.split('').slice(1).join('')
@@ -18,14 +18,4 @@ export const convertHEXToRGB = (hex: string) => {
   return rgb
 }
 
-export const toPascalCase = (value: string) => {
-  const valueList = value.split(/[-_\s]+/) // 구분자: 하이픈, 언더스코어, 공백
-
-  let result = ''
-
-  valueList.forEach((v) => {
-    result += v.charAt(0).toUpperCase() + v.slice(1).toLowerCase()
-  })
-
-  return result
-}
+export default convertHEXToRGB
