+++ src/shared/_utils/getElementById.ts
@@ -2,8 +2,10 @@ const isBrowser = () => {
   return typeof window !== 'undefined'
 }
 
-export const getElementById = (id: string) => {
+const getElementById = (id: string) => {
   if (!isBrowser()) return
 
   return document.getElementById(id)
 }
+
+export default getElementById
