+++ src/shared/_utils/getRandomNumber.ts
@@ -1,0 +1,5 @@
+const getRandomNumber = (min = 0, max = 1) => {
+  return Math.random() * (max - min) + min
+}
+
+export default getRandomNumber
