--- src/shared/_utils/number.ts
@@ -1,3 +1,0 @@
-export function getRandomNumber(min = 0, max = 1) {
-  return Math.random() * (max - min) + min
-}
